[BE] RENAME: 기존 log 폴더에 있던 dto를 admin/dto 폴더로 이동
